+++ src/main/java/com/macmie/mfoodyex/Controller/CreditCardMfoodyController.java
@@ -16,6 +16,12 @@ import java.util.List;
 
 import static com.macmie.mfoodyex.Constant.ViewConstant.*;
 
+/*
+ * be used when the requested resource cannot be found (null): HttpStatus.NOT_FOUND (404)
+ * be used when a successful request returns no content (empty): HttpStatus.NO_CONTENT (204)
+ * be used when the request is invalid or contains incorrect parameters: HttpStatus.BAD_REQUEST (400)
+ * */
+
 @RestController // = @ResponseBody + @Controller
 @RequestMapping(CREDIT_CARD_MFOODY)
 public class CreditCardMfoodyController {
@@ -26,39 +32,57 @@ public class CreditCardMfoodyController {
     private UserMfoodyInterfaceService userMfoodyInterfaceService;
 
     @GetMapping(URL_GET_ALL)
-    public ResponseEntity<List<CreditCardMfoody>> getAllCreditCardMfoodys(){
+    public ResponseEntity<?> getAllCreditCardMfoodys() {
         List<CreditCardMfoody> creditCardMfoodyList = creditCardMfoodyInterfaceService.getListCreditCardMfoodys();
-        if(creditCardMfoodyList.isEmpty()){
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        if (creditCardMfoodyList.isEmpty()) {
+            return new ResponseEntity<>("NO_CONTENT List of CreditCards", HttpStatus.NO_CONTENT);
         }
         return new ResponseEntity<>(creditCardMfoodyList, HttpStatus.OK);
     }
 
     @GetMapping(URL_GET_BY_ID)
-    public ResponseEntity<CreditCardMfoody> getCreditCardMfoodyByID(@PathVariable("ID") int ID){
+    public ResponseEntity<?> getCreditCardMfoodyByID(@PathVariable("ID") int ID) {
         CreditCardMfoody creditCardMfoody = creditCardMfoodyInterfaceService.getCreditCardMfoodyByID(ID);
-        if(creditCardMfoody == null){
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        if (creditCardMfoody == null) {
+            return new ResponseEntity<>("NOT_FOUND CreditCardMfoody with ID: " + ID, HttpStatus.NOT_FOUND);
         }
         return new ResponseEntity<>(creditCardMfoody, HttpStatus.OK);
     }
 
+    @GetMapping(URL_GET_BY_ID_USER)
+    public ResponseEntity<?> getCreditCardMfoodyByIdUser(@PathVariable("ID") int ID) {
+        if (userMfoodyInterfaceService.getUserMfoodyByID(ID) == null) {
+            return new ResponseEntity<>("NOT_FOUND UserMfoody with idUser: " + ID, HttpStatus.NOT_FOUND);
+        }
+        List<CreditCardMfoody> creditCardMfoodyList = creditCardMfoodyInterfaceService.getListCreditCardMfoodysByIdUser(ID);
+        if (creditCardMfoodyList.isEmpty()) {
+            return new ResponseEntity<>("NO_CONTENT CreditCard of UserMfoody with ID: " + ID, HttpStatus.NO_CONTENT);
+        }
+        return new ResponseEntity<>(creditCardMfoodyList, HttpStatus.OK);
+    }
+
     @DeleteMapping(URL_DELETE)
-    public ResponseEntity<?> deleteCreditCardMfoodyByID(@PathVariable("ID") int ID){
+    public ResponseEntity<?> deleteCreditCardMfoodyByID(@PathVariable("ID") int ID) {
+        if (creditCardMfoodyInterfaceService.getCreditCardMfoodyByID(ID) == null) {
+            return new ResponseEntity<>("NOT_FOUND CreditCardMfoody with ID: " + ID, HttpStatus.NOT_FOUND);
+        }
         creditCardMfoodyInterfaceService.deleteCreditCardMfoodyByID(ID);
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @DeleteMapping(URL_DELETE_BY_ID_USER)
-    public ResponseEntity<?> deleteCreditCardMfoodyByIdUser(@PathVariable("ID") int ID){
+    public ResponseEntity<?> deleteCreditCardMfoodyByIdUser(@PathVariable("ID") int ID) {
+        if (userMfoodyInterfaceService.getUserMfoodyByID(ID) == null) {
+            return new ResponseEntity<>("NOT_FOUND UserMfoody with ID: " + ID, HttpStatus.NOT_FOUND);
+        }
         creditCardMfoodyInterfaceService.deleteAllCreditCardsMfoodyByIdUser(ID);
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @PutMapping(URL_EDIT) // idUser is ignored
-    public ResponseEntity<?> editCreditCardMfoody(@RequestBody String creditCardPOJOJsonObject, BindingResult errors){
+    public ResponseEntity<?> editCreditCardMfoody(@RequestBody String creditCardPOJOJsonObject, BindingResult errors) {
         // Check Error
-        if(errors.hasErrors()){
+        if (errors.hasErrors()) {
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
 
@@ -67,8 +91,8 @@ public class CreditCardMfoodyController {
         CreditCardMfoodyPOJO newCreditCardPOJO = gson.fromJson(creditCardPOJOJsonObject, CreditCardMfoodyPOJO.class);
         CreditCardMfoody newCreditCardMfoody = newCreditCardPOJO.renderCreditCardMfoody();
         UserMfoody attachUserMfoody = userMfoodyInterfaceService.getUserMfoodyByIdCard(newCreditCardPOJO.getIdCard());
-        if(attachUserMfoody == null){
-            return new ResponseEntity<>("Can't find any UserMfoody with IdCard: " + newCreditCardPOJO.getIdCard(), HttpStatus.NOT_FOUND);
+        if (attachUserMfoody == null) {
+            return new ResponseEntity<>("NOT_FOUND UserMfoody with IdCard: " + newCreditCardPOJO.getIdCard(), HttpStatus.NOT_FOUND);
         }
         newCreditCardMfoody.setUser(attachUserMfoody);
 
@@ -78,9 +102,9 @@ public class CreditCardMfoodyController {
     }
 
     @PostMapping(URL_ADD) // idUser must be accurate
-    public ResponseEntity<?> addNewCreditCardMfoody(@RequestBody String creditCardPOJOJsonObject, BindingResult errors){
+    public ResponseEntity<?> addNewCreditCardMfoody(@RequestBody String creditCardPOJOJsonObject, BindingResult errors) {
         // Check Error
-        if(errors.hasErrors()){
+        if (errors.hasErrors()) {
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
 
@@ -92,13 +116,13 @@ public class CreditCardMfoodyController {
         // Check duplicate
         CreditCardMfoody existingCard = creditCardMfoodyInterfaceService.getCreditCardMfoodyByNumberCard(newCreditCardPOJO.getNumberCard());
         if (existingCard != null) {
-            return new ResponseEntity<>("A card with the same number already exists!", HttpStatus.CONFLICT);
+            return new ResponseEntity<>("CONFLICT - A card with the same number already exists!", HttpStatus.CONFLICT);
         }
 
         // Check input idUser and attach UserMfoody to CreditCardMfoody
         UserMfoody attachUserMfoody = userMfoodyInterfaceService.getUserMfoodyByID(newCreditCardPOJO.getIdUser());
-        if(attachUserMfoody == null){
-            return new ResponseEntity<>("Can't find any UserMfoody with ID: " + newCreditCardPOJO.getIdUser(), HttpStatus.NOT_FOUND);
+        if (attachUserMfoody == null) {
+            return new ResponseEntity<>("NOT_FOUND UserMfoody with ID: " + newCreditCardPOJO.getIdUser(), HttpStatus.NOT_FOUND);
         }
         newCreditCardMfoody.setUser(attachUserMfoody);
