+++ src/main/java/com/macmie/mfoodyex/Controller/FeedbackMailController.java
@@ -1,23 +1,23 @@
 package com.macmie.mfoodyex.Controller;
 
 import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
 import com.macmie.mfoodyex.Model.FeedbackMail;
-import com.macmie.mfoodyex.POJO.FeedbackMailPOJO;
 import com.macmie.mfoodyex.Service.InterfaceService.FeedbackMailInterfaceService;
-import com.macmie.mfoodyex.Util.CheckNullAPIRequest;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindingResult;
-import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
-import java.util.Map;
 
 import static com.macmie.mfoodyex.Constant.ViewConstant.*;
+
+/*
+ * be used when the requested resource cannot be found (null): HttpStatus.NOT_FOUND (404)
+ * be used when a successful request returns no content (empty): HttpStatus.NO_CONTENT (204)
+ * be used when the request is invalid or contains incorrect parameters: HttpStatus.BAD_REQUEST (400)
+ * */
 
 @RestController // = @ResponseBody + @Controller
 @RequestMapping(FEEDBACK_MAIL)
@@ -26,35 +26,36 @@ public class FeedbackMailController {
     private FeedbackMailInterfaceService feedbackMailInterfaceService;
 
     @GetMapping(URL_GET_ALL)
-    public ResponseEntity<?> getAllFeedbackMails(){
-        Gson gson = new GsonBuilder().setPrettyPrinting().create();
+    public ResponseEntity<?> getAllFeedbackMails() {
         List<FeedbackMail> feedbackMailList = feedbackMailInterfaceService.getListFeedbackMails();
-        if(feedbackMailList.isEmpty()){
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        if (feedbackMailList.isEmpty()) {
+            return new ResponseEntity<>("NO_CONTENT List of FeedbackMails", HttpStatus.NO_CONTENT);
         }
-        return new ResponseEntity<>(gson.toJson(feedbackMailList), HttpStatus.OK);
+        return new ResponseEntity<>(feedbackMailList, HttpStatus.OK);
     }
 
     @GetMapping(URL_GET_BY_ID)
-    public ResponseEntity<?> getFeedbackMailByID(@PathVariable("ID") int ID){
-        Gson gson = new GsonBuilder().setPrettyPrinting().create();
+    public ResponseEntity<?> getFeedbackMailByID(@PathVariable("ID") int ID) {
         FeedbackMail feedbackMail = feedbackMailInterfaceService.getFeedbackMailByID(ID);
-        if(feedbackMail == null){
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        if (feedbackMail == null) {
+            return new ResponseEntity<>("NOT_FOUND FeedbackMail with ID: " + ID, HttpStatus.NOT_FOUND);
         }
-        return new ResponseEntity<>(gson.toJson(feedbackMail), HttpStatus.OK);
+        return new ResponseEntity<>(feedbackMail, HttpStatus.OK);
     }
 
     @DeleteMapping(URL_DELETE)
-    public ResponseEntity<?> deleteFeedbackMailByID(@PathVariable("ID") int ID){
+    public ResponseEntity<?> deleteFeedbackMailByID(@PathVariable("ID") int ID) {
+        if (feedbackMailInterfaceService.getFeedbackMailByID(ID) == null) {
+            return new ResponseEntity<>("NOT_FOUND FeedbackMail with ID: " + ID, HttpStatus.NOT_FOUND);
+        }
         feedbackMailInterfaceService.deleteFeedbackMailByID(ID);
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @PutMapping(URL_EDIT)
-    public ResponseEntity<?> editFeedbackMail(@RequestBody String feedbackMailJsonObject, BindingResult errors){
+    public ResponseEntity<?> editFeedbackMail(@RequestBody String feedbackMailJsonObject, BindingResult errors) {
         // Check Error
-        if(errors.hasErrors()){
+        if (errors.hasErrors()) {
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
 
@@ -65,9 +66,9 @@ public class FeedbackMailController {
     }
 
     @PostMapping(URL_ADD)
-    public ResponseEntity<?> addNewFeedbackMail(@RequestBody String feedbackMailJsonObject, BindingResult errors){
+    public ResponseEntity<?> addNewFeedbackMail(@RequestBody String feedbackMailJsonObject, BindingResult errors) {
         // Check Error
-        if(errors.hasErrors()){
+        if (errors.hasErrors()) {
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
 
@@ -77,7 +78,7 @@ public class FeedbackMailController {
         return new ResponseEntity<>(newFeedbackMail, HttpStatus.CREATED);
     }
 
-    public FeedbackMail convertJsonToFeedbackMail(String feedbackMailJsonObject){
+    public FeedbackMail convertJsonToFeedbackMail(String feedbackMailJsonObject) {
         Gson gson = new Gson();
         return gson.fromJson(feedbackMailJsonObject, FeedbackMail.class);
     }
